uintptr_t

+++ ft_printf/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lin <lin@student.42.fr>                    +#+  +:+       +#+        */
+/*   By: linliu <linliu@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/06 09:51:27 by linliu            #+#    #+#             */
-/*   Updated: 2025/05/11 23:16:55 by lin              ###   ########.fr       */
+/*   Updated: 2025/05/13 14:16:08 by linliu           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <stdarg.h>
 # include <unistd.h>
+# include <stdint.h>
 
 # define HEX_LOWER "0123456789abcdef"
 # define HEX_UPPER "0123456789ABCDEF"
@@ -23,7 +24,7 @@ int	ft_printf(const char *format, ...);
 int	write_char(char c);
 int	write_str(const char *s);
 int	write_int(long n);
-int	write_uint_base(unsigned long n, const char *hex, unsigned long base);
+int	write_uint_base(uintptr_t n, const char *hex, uintptr_t base);
 int	write_pointer(void *ptr, const char *hex);
 
 #endif
