+++ ft_printf/write_uint_base.c
@@ -6,13 +6,13 @@
 /*   By: linliu <linliu@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/11 11:01:44 by linliu            #+#    #+#             */
-/*   Updated: 2025/05/12 09:42:09 by linliu           ###   ########.fr       */
+/*   Updated: 2025/05/13 14:16:34 by linliu           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	write_uint_base(unsigned long n, const char *hex, unsigned long base)
+int	write_uint_base(uintptr_t n, const char *hex, uintptr_t base)
 {
 	int	count;
 	int	check;
